+++ src/components/Cards/Bio.jsx
@@ -1,0 +1,16 @@
+export default function Bio() {
+
+    return <div className="card">
+
+        <div className="thumb" style={ { backgroundImage: 'url(userpic.gif)' } }/>
+
+        <article>
+            <h1>Maksim Korinets</h1>
+
+            <span>freelance software engineer</span>
+
+            <p>Embrace simplicity: there is hard work behind it (most of the time).</p>
+        </article>
+
+    </div>
+}