+++ tailwind.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  content: [
+    "./src/**/*.{js,jsx}",
+  ],
+  darkMode: false, // always dark mode
+}